test: add intent clusterer coverage
